// src/app/@core/modules/restaurant-and-finance-and-employee/model/TerminalTo.ts
export class TerminalTo {
    dateFrom: Date;
    dateTo: Date;
    cashMoney: number;
    cardMoney: number;
    bankTransferMoney: number;
    deliveriesOrdersQuantity: number;
    deliveriesMoney: number;
    takeawayOrdersQuantity: number;
    takeawayMoney: number;
    localOrdersQuantity: number;
    localMoney: number;
}